Update Gaussian_mixture_of_proximity_points.py
The code provides two methods for label enhancement.The first method is based on ordinary distance weighting, and the second method is based on the linear representation of popular learning.These functions will  return a dict, 'mean' represents the means of the mixed Gaussian, and 'weight' represents the weight of the different individual Gaussian.The code also provides two methods for variance. 'var_settled' represents fixed variance, and 'var_determined_by_weight' represents variance adjustment according to weight. The code also provides a function for plotting the mixed Gaussian.